Działa, jest ok.

+++ lab1/inc/doubleNumbers.h
@@ -1,14 +1,16 @@
 #ifndef DOUBLENUMBERS_H_
 #define DOUBLENUMBERS_H_
 
+#include <vector>
+
 class doubleNumbers {
 private:
-	int _size;
-	long* _numbers;
+	unsigned long long _size;
+	unsigned long long* _numbers;
 public:
 	doubleNumbers();
-	doubleNumbers(int size);
-	void multiplyByTwo(int number);
+	doubleNumbers(unsigned long long size);
+	void multiplyByTwo(unsigned long long number);
 	void loadRandomNumbers();
 };
